backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
youtube-transcript-api
requests
faster-whisper